Apply the updates in transaction
Ideally, when we make changes to the DB they should all succeed or all fail.

We've not been too great at doing that with notices and notifications, but this seemed a small step we could make when it comes to updating the notification after checking the status, plus any dependent records.