Add 'userRoles' function to the base licence presenter. (#2671)
* Add 'userRoles' function to the base licence presenter.

https://eaflood.atlassian.net/browse/WATER-5372

We have moved from using tabs to 'secondary navigation'. This means the navigation of the page is now on the left-hand side.

We need to remove the use of the common data as this is no longer used across licence and is only needed in the licence summary page.

This change lifts the common data 'roles' logic into its own helper function for all the licence pages to use instead of the 'commonData'