Set sent date on rtn logs when confirmed as sent (#2670)
https://eaflood.atlassian.net/browse/WATER-5019

We've previously made changes to our returns notice engine to capture the return log IDs linked to a recipient.

For example, when creating a returns invitation, we start with the 'due' return logs for the selected period. They give us the licence references from which we determine the recipients for the invitation.

Previously, once we had the licence references, the return logs no longer mattered. Knowing this change was coming, we updated the query to return an array of the return log IDs for each recipient, populated by those that determined the licence refs, which then determined the recipient. These arrays of return log IDs are then persisted against each notification.

As a result, we can now implement this change. Once we receive confirmation that a notification has been 'sent' by Notify, we will set the sent date on those return logs linked to the notification.

That is after we make `sent_date` accessible in the view!

The sent date was previously populated by the NALD and legacy notice engines. It wasn't critical, but we're implementing dynamic due dates to support 'Special measures'. It is about ensuring returns are submitted in accordance with the licence within 28 days of being notified that they are required.

This is why return log due dates now need to be tied to when the notification is sent, instead of 28 days after the cycle ends.

Continuing to record _when_ a return log was sent will help our compliance-checking tools identify issues and when they occur.