Housekeeping notice setup page & process services
https://eaflood.atlassian.net/browse/WATER-4991

The next element of dynamic due dates we need to implement is a change to the query when fetching applicable return logs for an ad-hoc returns invitation.

It means making an _another_ change to the fetch services, so rather than making them yet more complex we're intending to split them out.

But when we were about to start doing this, we realised that first we'd need to do a bit of housekeeping with the existing services.

It's not clear when a service is a 'service', i.e. functionality in support of a page or process, for example, `FetchRecipientsService`. Or where it is just to handle viewing a page or processing a request.

Also, only some things were placed into folders, for example, `returns-period/` service, but the majority were not.

So, before we start splitting up some of our existing 'service' services, we want to do some housekeeping to avoid some the confusion and inconsistency already present.