Update licence bill page heading (#2665)
* Update licence bill page heading

https://eaflood.atlassian.net/browse/WATER-5372

We have moved from using tabs to 'secondary navigation'. This means the navigation of the page is now on the left-hand side.

We need to remove the use of the common data as this is no longer used across licence and is only needed in the licence summary page. All other pages only need to fetch the licence id and licence ref.

The page headings are being updated so the caption is the licence ref and the page title what the page is, in this case 'bills'.

This change decouples the licence bills page from the 'commonData' and fetches the licence id and licence ref within its own fetch query.

This pattern is repeated across the updated licence pages.