Rename & move view & submit pages
The next bit we're tidying up is the name and location of view-based services.

By prefixing the names with `view-` it makes it easier to match to their respective controller and distinguish them from non-page based services.

At one stage with the notice engine we were going to separate services by page or type, but this hasn't been applied consistently. So, we're moving all view-page based services to the same level.

From there we'll have a clearer idea of how best things should be organised.