Rename non-view and submit endpoints
At this point its clear which services in `app/services/notices/set-up` handle pages, whether its viewing or submitting.

But we're still left with a plethora of services, some of which support these pages, others which like them, are handling requests to an end point.

In our jobs, we have a convention of prefixing the service that is called by the controller `process-`. This seems a great way to delineate those services that handle requests, but are not responsible for returning page data for a view, or handling a form submission. Instead, they perform some 'process' when the request is received.

So, in the last series of changes, we're going to rename those notice-setup services that deal with processes, triggered from a request.