Update licence sub navigation pages to updated layout (#2680)
* Update licence sub navigation pages to updated layout

https://eaflood.atlassian.net/browse/WATER-5372

We have moved from using tabs to 'secondary navigation'.

This means the navigation of the page is now on the left-hand side, and the conditions / points and purposes are in a sub navigation under the primary 'Licence summary' navigation.

This change updates the layouts of the pages to use the licence layout.

The links follow the existing pattern of relative urls in the layout. So 'points' will take the user on the current path + 'points' e.g. '/system/licences/{id}/points'

We have introduced 'activeSummarySubNav' to toggle the aria current in the view. This works in combination with 'activeTab' to only show the sub navigation on the licence summary grouped pages.

'activeTab' is marked to be updated in another change.