Housekeeping notice setup page & process services (#2682)
https://eaflood.atlassian.net/browse/WATER-4991

The next element of dynamic due dates we need to implement is a change to the query used to fetch applicable return logs for an ad-hoc returns invitation.

It means making an _another_ change to the fetch services, so rather than making them yet more complex we're intending to split them out.

But when we were about to start, we realised we'd first need to do a bit of housekeeping with the existing services.

It's not clear when a service is a 'service', i.e. functionality in support of a page or process, for example, `FetchRecipientsService`. Or is it just for handling page views or processing requests?

Also, only some things were placed into folders, for example, `returns-period/` service, but the majority were not.

So, before we start splitting up some of our existing 'service' services, we want to do some housekeeping to avoid some of the confusion and inconsistency already present.