Update licence summary to use secondary navigation (#2649)
https://eaflood.atlassian.net/browse/WATER-5168

This change is a significant update to the layout of the view licence pages. We are moving from [tabs](https://design-system.service.gov.uk/components/tabs/) to [secondary navigation](https://govuk-prototype-components.x-govuk.org/secondary-navigation). This means the page navigation is now on the left-hand side.

The requirement is that all pages accessed from the navigation menu should also display the navigation menu.

Plus, from this point on, details such as the licence name, whether it is registered, and any warnings or notifications would only be displayed on the licence summary page, rather than every page. The rest of the pages will use a simplified title and the selected licence reference as the caption.

Before this change, the view licence tab pages were informally a mix of two views. We used this opportunity to create a layout specific to these pages formally, implement the second navigation menu, and ensure other details remained consistent.

The output of this layout is that the secondary navigation takes up one quarter, and the 'main' content for the pages takes up the other three quarters.

There was also a requirement to make the 'Skip to main' go to the specific page data, rather than the first secondary navigation option. The only solution to support this was to spread the divs across the `beforeContent`, `content` and `afterContent` blocks in the layout.

It was easier to do this en masse rather than individually, so the change encompasses several tickets.

- [WATER-5372 Add sub-navigation under Licence summary link](https://eaflood.atlassian.net/browse/WATER-5372)
- [WATER-5170 Switch to secondary navigation on Licence summary page](https://eaflood.atlassian.net/browse/WATER-5170)
- [WATER-5176 Switch to secondary navigation on Contact details page](https://eaflood.atlassian.net/browse/WATER-5176)
- [WATER-5177 Switch to secondary navigation on Returns page](https://eaflood.atlassian.net/browse/WATER-5177)
- [WATER-5178 Switch to secondary navigation on Communications page](https://eaflood.atlassian.net/browse/WATER-5178)
- [WATER-5179 Switch to secondary navigation on Bills page](https://eaflood.atlassian.net/browse/WATER-5179)
- [WATER-5180 Switch to secondary navigation on Licence set up page](https://eaflood.atlassian.net/browse/WATER-5180)
- [WATER-5181 Add secondary navigation to Purposes, periods and amounts page](https://eaflood.atlassian.net/browse/WATER-5181)
- [WATER-5182 Add secondary navigation to Points page](https://eaflood.atlassian.net/browse/WATER-5182)
- [WATER-5183 Add secondary navigation to Conditions page](https://eaflood.atlassian.net/browse/WATER-5183)
- [WATER-5169 Display only title and licence reference as caption](https://eaflood.atlassian.net/browse/WATER-5169)