Update licence counts to pluralise summary counts
https://eaflood.atlassian.net/browse/WATER-5372

When we show points / conditions / purposes, we need to pluralise the word whe the total items are 1.

This change introduces a 'pluralise' helper function to pluralise a work based on the count provided.

The points / conditions / purposes presenters have been updated to use this pluralise function for the showing counts sentence.