Update fetch licence ref to fetch licence (#2685)
https://eaflood.atlassian.net/browse/WATER-5372

Previously we had a 'FetchLicenceService' that was used for a specific case. This has been refactored to something else.

This change renames the 'FetchLicenceRefService' to the 'FetchLicenceService'.

This will be used by all the licence pages to get the id and licence ref. (changes to make in later changes)

Some pages use the '$ends()' helper and so the required dates are provided for this functionality.