Move to notice-type specific fetch services
https://eaflood.atlassian.net/browse/WATER-4991

The next element of dynamic due dates we need to implement is a change to the query when fetching applicable return logs for an ad-hoc returns invitation.

In the standard journey _only_ those with no due date that match the period selected are used to determine the recipients.

Currently for ad-hoc returns invitations, we don't care about the period (as long as we are past a return logs end date), but they still have to have a NULL due date.

This will mean yet another scenario and change to the fetch services for returns notifications, that then has to be applied to the remove licences query and the download query.

We're adding a lot of complexity to those services, to avoid duplicating the some static text (the stuff which actually grabs and groups the recipients once the return logs have been selected).

It's also hard to trace what query will be applied in what scenario.

Because of this, before making the change for ad-hoc return invites we're going to re-organise what we have, splitting up the fetch services.

We're starting this PR with the intent of splitting them by notice type but we'll see what works best as we refactor!