Rename and update FetchReturnsDueService
The service was fetching return logs that were due for the period and notice type selected. But each result only contained the licence ref of the return log, and any duplicates were removed.

All we really wanted were the licence refs from return logs due in the period selected (plus matching cycle, siCurrent, and due date based on notice type).

Prior to this change the 'ReturnLogModel' instances were then being passed to the validator, where they were then turned into an array of licence references.

So, amongst of fetch services the service name was confusing, and it wasn't really returning the result we wanted. Plus, it was made slightly more complex by worrying about trying to filter it by whatever licence refs had been entered for removal.

Simpler, and the intent is clearer, to not worry about that aspect. The performance benefit is negligible to none existent so better to simplify our service.