Update title
Now we're removing notices from the manage page we can't have it in the title.

But the existing title not matching the nav bar is inconsistent with the other pages accessed from the nav bar. So, rather than change to something like 'Manage reports and users', we change it to just 'Manage'.