Remove support & links 4 notices from manage page (#2693)
https://eaflood.atlassian.net/browse/WATER-5404

With the rebuild of sending notices, we brought creating a standard or ad-hoc notice into the 'Notices' page, [promoted it to the nav bar](https://github.com/DEFRA/water-abstraction-system/pull/2553), and removed links to redundant notice types. Essentially, 'Notices' is now a feature in its own right, like Bill Runs, instead of a 'report' and a bunch of separate links.

Rather than perform a 'hard-switch', we continued to support the old links on the manage page into the new notice journey. But this is unnecessary complexity to support, both in terms of the code and in requiring users to understand that there are two ways to do the same thing.

This change removes the links from the 'Manage' page and updates the title to be consistent with all the other pages accessed from the nav bar.

It then drops the support in the notice setup journey to know from which point the journey was started. Now it will always be the 'Notices' page.