Move generateReferenceCode to general lib
`DetermineNoticeTypeService` is actually called at different times by the difference notice types, for example, abstraction alerts when the session is initiated, and the rest when the notice type is selected.

Other than the notice lookup, the only thing is does is generate the reference code. However, we also have a copy of a pseudo reference code generator in the notification helper.

If we move the code generation to the lib, we can test it in isolation, and re-use it in our tests.

Then, the only thing `DetermineNoticeTypeService` is left to do is a lookup on an object, something which can easily be done by the services that need this info, resulting in one more service we can make redundant.