Send letter returns invitation after email failure (#2664)
https://eaflood.atlassian.net/browse/WATER-5230

As part of our work on setting dynamic due dates for return logs we want to be certain that when we send out invitations to submit return logs that the notification has been received. If the account has an email address associated with it we will use that in the first instance. This could be to a returns agent or registered user. If for what ever reason that email fails to send (be it an incorrect email address or a technical issue on our or GovNotify's end) we will then send a letter to the licence holder.

This PR adds several services and incorporates the logic at the end of the journey for creating the returns invitation. We know by the end if the email was successful or not so we can then create the new notice for a letter invitation.