Update sub nav pages for common licence fetch (#2695)
* Update sub nav pages for common licence fetch

https://eaflood.atlassian.net/browse/WATER-5376

We need to update all of our licence pages to use the common fetch licence service within their own services.

To achieve this we need to update the current services to be in the 'shape' required for the refactor. That is the fetch's return an object containing the licence and the relevant data e.g { licence, points }.

When we refactor we rip out the licence from all the fetches.

This change updates the points / conditions and purposes to meet this shape.