Migrate licence holder search (#2667)
https://eaflood.atlassian.net/browse/WATER-5399

We are migrating the internal search page (the root page of the internal service) from the legacy apps to our new `water-abstraction-system` app.

The existing page allows users to search using different types of values, for example, licence references, licence names, and return references.

We are implementing these different types of search into the new `system` page one at a time to manage the complexity.

Functionality to search for licences, returns, monitoring stations, users and billing accounts has already been implemented. This change is to implement searching for licence holders.

This is the last search item for now, but will be followed by another change that will rework how the database querying is done.