Add default No x for this licence to licence pages.
https://eaflood.atlassian.net/browse/WATER-5376

Some of the licence pages show a message of 'No bills sent for this licence.' This covers situations when the specific page information is empty, in this case bills.

The conditions, points and purposes did not have this default as preciously the links where on the summary page.

As they are now in the secondary navigation under the licence summary and sub navigation we need to add these default messages.

This change adds the default message 'No X sent for this licence.' to the relevant pages.