Update the licence model current versions (#2708)
* Update the licence model current versions

https://eaflood.atlassian.net/browse/WATER-5376

We have identified an issue with the current logic to work out the licence version current version.

A licence version can be marked as 'current' but the start date could be in the future (this comes from NALD).

We can no longer rely on the status field and the status of 'current' to return the current version.

This change updates the current version logic to get the 'current' licence version by its start date.

If the date is <= today, it is a valid 'current' licence version. The latest of these start dates is the 'current' licence version.

We use the modifier query to order the licence versions and return the latest (<= today).