Search using single union  query
https://eaflood.atlassian.net/browse/WATER-5292

We are migrating the internal search page (the root page of the internal service) from the legacy apps to our new `water-abstraction-system` app.

Each different type of search has been migrated individually so far.

This change is to combine then into a single, simpler query that will display all the results in a logical sequence, correctly paginated.