Update licence contacts (simplify)
https://eaflood.atlassian.net/browse/WATER-5376

We have two contact pages which use different data sources. There has been some confusion on the code base of which one it which.

This change updates the contacts to reflect the recent refactoring of the licence pages.

A contact and customer contact are shown as a licence page under the 'Contact details' tab. Theses have been updated to simply be 'contact' and 'customer contact'.

A licence contact is a contact shown on the 'licence-contact' page.

The previous states have been ignored from the commit to ignore any further confusion.