Add a notification banner to the licence-set-up page
https://eaflood.atlassian.net/browse/WATER-5376

We recently refactored the licence pages to use a secondary navigation. As part of this change, we only wanted to show the notification on the licence summary page.

This now needs to be update and we need to show the notification on the licence set up page.

This change adds the required data to the fetch licence, lifts the notification logic out of the licence summary presenter into the base licence presenter, and sets the notification for the set up page.