Properly sort out names for contact pages
In this change we ensure everything from the route to the presenter and associated tests are aligned.

Essentially, the page accessed by clicking a link _within_ the summary content goes to `/licence-contacts` (we add the 's' in this change to tidy things up). All its modules and references have dropped the word 'details', including from the title just to make this point perfectly clear, and the page distinct.

The page accessed from the secondary navigation menu that goes to `/contact-details`, is now the only place you'll find reference to 'contact-details', both in its module names and content.

We also fix the fact that the test files for the presenters were looking at the other presenter! Proof this needed sorting.