Tidy up view licence file naming for consistency (#2711)
https://eaflood.atlassian.net/browse/WATER-5168

This is a 'tidy' up of the routes, services, presenters and tests related to viewing a licence.

The most egregious issue we had to deal with was 'contacts'. What was once a page and a tab is now two distinct pages, with similar titles and, for most licences, the same content.

Both previously had the words 'contact' and 'details' in their titles, leading to a proper cufuddle in our code!

Now we've made it clear which code relates to which page by only allowing one of them to use the word 'details'. Whilst doing this, we even spotted that the test presenter files were testing the other pages' presenters! Proof that this needed sorting.

Next was ensuring that the 'things' align. For example, viewing a licence's returns. It will be on the route `/licences/{id}/returns,` so the file names should align as

- returns.presenter.js
- fetch-returns.service.js
- view-returns.service.js
- returns.njk

We went through what we have and ensured this. This involved renaming the summary page's presenters so the main content is now handled by `summary.presenter.js` to match the other pages.