Add the licence history page into secondary navigation (#2714)
* Add the licence history page into secondary navigation

https://eaflood.atlassian.net/browse/WATER-5071

We recently implemented an improved layout to the licence pages using secondary navigation.

The History needs updating the conventions set in the other licence pages (using the secondary navigation layout).

This change updates the current history page to work with the secondary navigation layout.

The page content has remained the same (bar the headings).

This change also introduces a feature toggle for the 'history' secondary navigation action.