Update licence history to only licence versions (#2716)
* Update licence history to only licence versions

https://eaflood.atlassian.net/browse/WATER-5071

The updated history page only shows licence versions. So we need to remove the returns and charges from the fetch query and the view.

This change strips the history page back to showing only the licence versions.

It only shows the start and end date in the view (and the view action button).

The UI has been updated to use the updated table, the other fields have been left null and will be implemented in a later change