Fix Determine Returns Period service failing test
We have a consistently failing test where our stub of the 'clock' is being reset.

This test was using a before to set the clock and has been working with no issues up until now.

It is not clear why this has started to fail. But it is.

This change updates setting the clock before each test rather than before the subtest suite is run. This appears to 'fix' the issue.