Update licence history to use 'reason' (#2718)
https://eaflood.atlassian.net/browse/WATER-5071

We recently refactored the history page to only return show the licence versions. We left out the reason to simplify previous changes (this is behind a toggle)

This change updates the fetch to return the 'modLogs' so we can show the 'reason' for each licence version. The presenter has been updated to display the reason using the '$reason()' helper on the 'LicenceVersionModel'.