Update licence history to use 'changeType' (#2722)
* Update licence history to use 'changeType'

https://eaflood.atlassian.net/browse/WATER-5071

We recently refactored the history page to only return show the licence versions. We left out the change types to simplify previous changes (this is behind a toggle)

This change updates the fetch to return an 'administrative' field, this field is used in the presenter and UI.

A Change type has 2 options - based on whether a version is an issue or increment. When it's an Increment ‘No licence issued’ will be used and When it's an Issue 'Licence issued' will be used.