Move to generate query services
After much back and forth with the code and how to split it up, we came to the conclusion that the problem we're dealing with is a series of very similar, but often slightly different queries to fetch the recipient data.

For example, we could create a fetch service for paper returns, but the only difference between it and a returns invitation is the exclusion of primary users and returns agents.

Another example, is that a paper returns and a failed invitation share the same `due_returnLogs` requirement: fetch by ID. However, the recipient bit is again slightly different as failed invitations _only_ want licence holders.

Also, the query itself has two core parts. The first is the query for the due return logs. This is the one that changes the most dependent on the notice type. The second is then the fetching and processing of the recipient data. That is essentially the same whether sending or downloading the data. If the fetch is for the download, the results are formatted different but the rest is the same.

Having a fetch per notice note results in lots of duplication, both in the code and the work to test the services.

So, having worked through all this, the result we've ended up with are a series of query generation services. We have 3 for the different return log queries we need to use, and then one for fetching the recipients.

Breaking them up like this means we can test each in isolation, and avoid duplication. Also, we can still have notice specific fetch services, but now they govern how the generate query services are used to fetch _their_ recipients.

The first commits are adding the new services. We can then start the replaced existing services. The final result will be a much slimmed down notice engine with clearer service paths.