Add GenerateRecipientsQueryService
This is the core service that handles how we fetch and process the recipient data for all the notice types.

We've brought all the different types together into one service, to make it easier to see how the query changes depending on the notice type, and whether we are sending/checking or downloading the recipients.

We've also tried to make it clearer the different recipient scenarios we have to deal with, and how the query results will change depending on the notice type.