Update scaffolding jsdocs (#2725)
It has been noticed in a PR that the jsdoc return block is no in the expected conventions.

We currently return have - '@returns {object} - The data formatted for the view template'. The '-' should not be used in a return block.

This change updates the scaffolding to remove the '-' from the return block.