Update recipients fixture
We update the fixture to be more flexible in the recipients being generated. Rather than assuming we will always want _all_ abstraction alert fixtures, or all non-alert alert fixtures, it allows us to pick the fixtures we need in our tests.

It also ensures the fixtures reflect the results we are now getting out of our new services.