Update licence version page to show reason and notes (B&D user)
https://eaflood.atlassian.net/browse/WATER-5382

The licence version pages needs to show the notes and reason for the licence versions. A billing and data user can see additional information on the page. This is 'cerated on' and 'created by'.

This change adds the reason and notes to the licence version page.

When there is no reason ('reasonDescription') we do not show the reason.

When there are no notes (there could be multiple), we do not show the notes (or preceding 'Notes:' text).

Non B&D users changes will be made in a separate change.