Remove redundant download presenters
Part of our changes to how recipients are fetched is to ensure that ad-hoc, standard and paper returns results all have the same format. Therefore, their CSV files will now have the same format.

This means we can use a single presenter for formatting the data for the download CSV file.