Move to notice-type specific fetch services (#2692)
https://eaflood.atlassian.net/browse/WATER-4991

The next element of dynamic due dates we need to implement is a change to the query when fetching applicable return logs for an ad-hoc returns invitation.

In the standard journey, _only_ those with no due date that fall within the selected period are used to determine the recipients.

Currently, for ad-hoc returns invitations, we don't care about the period (as long as we are past a return logs end date), but they still have to have a NULL due date.

This will mean yet another scenario and change to the fetch services for returns notifications, which then have to be applied to the remove licences query and the download query.

We're adding a lot of complexity to those services to avoid duplicating the same static text (the stuff which actually grabs and groups the recipients once the return logs have been selected).

It's also hard to trace which query will be applied in which scenario.

Because of this, we're going to reorganise what we have.

After much back-and-forth on the code and how to split it up, we concluded that the problem we're dealing with is a series of very similar, but often slightly different, queries to fetch recipient data.

For example, we could create a fetch service for paper returns. Still, the only difference between it and a returns invitation is the exclusion of primary users and returns agents (emails essentially).

Another example is that a paper returns and a failed invitation share the same `due_return_logs` requirement; to fetch them by ID. However, the recipient bit is again slightly different as failed invitations _only_ want licence holders.

Also, the query itself has two core parts. The first is the query for the `due_return_logs`. This is the one that changes the most, depending on the notice type. The second is the fetching and processing of the recipient data. That is the same whether sending or downloading the data. The only difference is how the results are formatted.

Having a fetch per notice note results in significant duplication, both in the code and the work to test them.

So, having worked through all this, we've ended up with a series of query-generation services for returns notices. We have three for the different return log queries we need to use, and then one for fetching the recipients.

Breaking them up like this lets us test each in isolation and avoid duplication. Also, we can still have notice-specific fetch services, but now they govern how the generate-query services are used to fetch _their_ recipients.

With all these changes, there was a lot of duplication and redundant services we could drop.

The result is hopefully a simpler notice engine, where the route from hitting the `/check` page to fetching the specific recipients for the notice, and the journey a user is on, is much clearer.