Add logic in submitLicence to set latestDueDate
We'll check this when 'null due dates' is enabled instead of the one in `determinedReturnsPeriod` in the `NotificationsPresenter`.

Once enabled the `dueDate` in `determinedReturnsPeriod` will always be null. This works fine for standard returns invitations and reminders because we always want to be calculating their due dates based on the current date.

But for ad-hoc, if all the 'due' return logs for the licence have a due date, we're supposed to use the latest one.

So, going forward the only time we won't be calculating the due date in the presenter is when this new property is populated.

> We are encurring some technical debt with this one. We need to get this change in the hands of QA. But the unit tests for the service really need to be re-written. We should be stubbing `FetchDueReturnsForLicenceService`, not creating data for it, for each and every test in the unit test file.