Determine due date when sending ad-hoc notices (#2729)
https://eaflood.atlassian.net/browse/WATER-5222

As part of dynamic due dates for return logs, we've been asked to change how the 'due date' is calculated for ad-hoc returns invitations and reminders.

Currently, it's 27 or 28 days from the current date. Now, we've been asked to look at all the 'due' return logs for the selected licence.

If any have a null 'due date', then we carry on as we are. However, if they all have a 'due date', we are to pick the latest one and use it in the notifications.