Void return logs after end date change (#2731)
https://eaflood.atlassian.net/browse/WATER-5403
https://eaflood.atlassian.net/browse/WATER-5362

Users have found that when they extend the end date of a licence we're not correctly voiding the previous return logs when we create the new one.

We were previously only voiding return logs after the change date rather than inclusive of the change date. This PR adds the change date into the void query.