Update licence version page to show reason and notes (#2727)
https://eaflood.atlassian.net/browse/WATER-5382
https://eaflood.atlassian.net/browse/WATER-5383

The licence version page needs to show the licence versions reason. This is show to all users.

When the user is for billing and data then we show some additional data in the reason, this is the created on and created by.

A billing and data user will also see the notes in a separate section.

This change adds the reason and notes to the licence version page, these are shown based on user permissions.

When there is no reason ('reasonDescription') we do not show the reason for a non billing and data user, if the user is for billing and data then we will always show the 'Created on' text.

For a billing and data user, when there are no notes (there could be multiple), we do not show the notes (or preceding 'Notes:' text).