Add error in data text to licence version page (#2732)
https://eaflood.atlassian.net/browse/WATER-5383

When a user is not a billing and data user we want to show some text with an email to allow the user to report data issues.

This changes adds conditional text to the licence version page. This text shows the user how to report a data issue, this has an email with a 'mailto' link.