Enable use of new search page
https://eaflood.atlassian.net/browse/WATER-5292

We are migrating the internal search page (the root page of the internal service) from the legacy apps to our new `water-abstraction-system` app.

The new search page has been created and can be enabled using a new feature flag `USE_SYSTEM_SEARCH` in the `water-abstraction-ui` application.

This flag tells the application to use the new search page instead of the old one. The page is served from the system application at `/system/search`.
The `water-abstraction-ui` application redirects any requests to its root to the search page - when the feature flag is set that will be the new search page.

To avoid multiple redirects, any links that point to the old  `/licences' path have been updated to point to the correct root path `/` instead.