Add next and previous presenter (#2733)
https://eaflood.atlassian.net/browse/WATER-5190

As part of the work to add the 'previous' and 'next' pagination links to the licence versions page we have decided to make a common component to support a common implementation for future work.

This change introduces the 'PaginatorPreviousAndNextPresenter', which will receive a sorted array and an id and will return the items either side of the ids index in the array.