Record contact type in notifications
https://eaflood.atlassian.net/browse/WATER-4991

This is actually related to our work on dynamic due dates. Going forward, due dates will be set against return logs when we get confirmation from Notify that the related notification was successful.

However, we should only be doing this for certain notice and contact types.

What we've realised is the logic we have for this is faulty, because it only has the `messageRef` to go on. For example, in an ad-hoc notice you can add a 'single use' recipient. Because we have to set something for `message_ref`, we've opted to go with `returns_invitation_primary_user` for email contacts, and `returns_invitation_letter-holder` for letter contacts. These were legacy values used when it handled notices.

But now we've 'lost' the knowledge that this was an 'single use' contact, not the licence holder or primary user. Plus, applying due dates to a return log because its a successful 'returns invitation'.

Ideally, the two values should be separate; we are sending a **returns invitation** to the **primary user**, or a **paper return** to a **single use** contact. Then decisions like, what notification decides whether due dates are applied or not, become much simpler and clearer to implement.

In this change we are going to start making this distinct clear.

- `message_ref` will be simplified to just the notice type
- a new field `contact_type` will be populated with the appropriate 'contact type'
- affected functionality such as viewing and filtering notices and notifications will be updated