Add previous/next pagination to licence version page
https://eaflood.atlassian.net/browse/WATER-5190

We recently implemented a common previous/next presenter to show the previous/next links for the pagination components.

This change adds the logic to show the previous / next licence version pagination links on the licence versions page.

These are different from our usual pagination in that we require knowing all the elements in the array to then pick the previous and next elements based on the currents id.

This required the fetch to return this array of elements. We might be adding this pagination to other pages, but because those pages will show the link text differently we have implemented the pagination as page specific.