Fix totals when viewing a return submission The totals for the user quantities entered were being calculated by adding together all the quantities in cubic metres, and then converting that total to the UOM used by the user.
This was causing problems when there were a lot of decimals involved as the total did not add up to the sum of the user unit quantities. To calculate this total properly the lines are converted from cubic metres to the users unit on a line by line bases. Then the total of all these lines is summed.