Round return submission entries to 6 DP (#2668)
https://eaflood.atlassian.net/browse/WATER-5397

When editing/creating a return submission ensure all values are rounded to 6 decimal places when persisted.

Ensure all values are displayed "as is" in the UI.

A major part of this change is how the quantities are now stored in the session and calculated. Previously, whatever the user entered as the quantity was just stored in the session "as is" regardless of the UOM. As the quantities are stored in the DB in cubic metres only, the conversion to cubic metres was done on the fly when viewing on the check page, and calculated prior to commiting the data to the DB. Once we started rounding to 6 decimal places this was causing inconsistencies in the totals when editing a return, particlary when gallons was used as the unit of measurement.

This inconsistency was occuring because the totals are only stored in cubic metres in the DB. Which meant when editing a return submission, when the quantities were being converted back to the original UOM used when the submission was first created. In some cases, particulary when gallons were being used. The quantities & totals when displayed in the original UOM were no longer matching what was originally entered.

As this rounding issue isn't something we can "fix" since its just a product of rounding to 6 DP. What we are now doing when a quantity is entered, either as a single volume or on individual lines. Is to take the quantity entered, convert it to cubic metres and store it in the session as `quantityCubicMetres`. Then take the value for `quantityCubicMetres` and convert it back to the UOM used by the user and save that value as the `quantity`. This ensures that the quantity in the UOM chosen by the user, when played back,  will always reflect the value that is actually being charged.

Some inset text has also been added to pages where volumes are entered to warn the user of this conversion. An additional benefit of now storing the values in the session in both the users UOM and cubic metres means that we can just display/save the quantites that are in the session rather than having to keep recalculating them.

Part of this work will also be ensuring consistency in the services that make up the return logs setup journey. As well sorting out the issue identified in https://github.com/DEFRA/water-abstraction-team/issues/155